+++ samples/mario-pizzeria/domain/entities/__init__.py
@@ -1,19 +1,24 @@
 """Domain entities for Mario's Pizzeria"""
 
 # Export all entities for clean import access
-from .enums import PizzaSize, OrderStatus
-from .pizza import Pizza
-from .customer import Customer
-from .order import Order
+from .customer import Customer, CustomerState
+from .enums import OrderStatus, PizzaSize
 from .kitchen import Kitchen
+from .order import Order, OrderState
+from .order_item import OrderItem
+from .pizza import Pizza, PizzaState
 
 __all__ = [
     # Enums
     "PizzaSize",
     "OrderStatus",
-    # Entities
+    # Entities & States
     "Pizza",
+    "PizzaState",
     "Customer",
+    "CustomerState",
     "Order",
+    "OrderState",
+    "OrderItem",  # Value object
     "Kitchen",
 ]